Amélioration de la structure du code 🎨
