+++ modeles/personnalite.dao.php
@@ -12,78 +12,6 @@ class PersonnaliteDAO
     }
 
     // Méthodes
-    function create(Personnalite $personnalite): bool
-    {
-        $req = $this->pdo->prepare("INSERT INTO " . DB_PREFIX . "personnalite (nom, prenom, urlImage) VALUES (:nom, :prenom, :urlImage)");
-        $nom = $personnalite->getNom();
-        $prenom = $personnalite->getPrenom();
-        $urlImage = $personnalite->getUrlImage();
-        
-        $req->bindParam(":nom", $nom);
-        $req->bindParam(":prenom", $prenom);
-        $req->bindParam(":urlImage", $urlImage);
-        
-        $success = $req->execute();
-        
-        if ($success) {
-            // Récupérer l'ID généré et le mettre dans l'objet
-            $id = $this->pdo->lastInsertId();
-            $personnalite->setId((int)$id);
-        }
-        
-        return $success;
-    }
-
-    function update(Personnalite $personnalite): bool
-    {
-        $req = $this->pdo->prepare("UPDATE " . DB_PREFIX . "personnalite SET nom = :nom, prenom = :prenom, urlImage = :urlImage WHERE id = :id");
-        $id = $personnalite->getId();
-        $nom = $personnalite->getNom();
-        $prenom = $personnalite->getPrenom();
-        $urlImage = $personnalite->getUrlImage();
-        
-        $req->bindParam(":id", $id);
-        $req->bindParam(":nom", $nom);
-        $req->bindParam(":prenom", $prenom);
-        $req->bindParam(":urlImage", $urlImage);
-        return $req->execute();
-    }
-
-    function delete(int $id): bool
-    {
-        $req = $this->pdo->prepare("DELETE FROM " . DB_PREFIX . "personnalite WHERE id = :id");
-        $req->bindParam(":id", $id);
-        return $req->execute();
-    }
-
-    public function hydrate($row): ?Personnalite {
-        if (!$row) {
-            return null;
-        }
-
-        // Récupération directe des valeurs depuis le tableau associatif
-        $id = $row['id'] ?? $row['idPersonaliter'] ?? null;
-        $nom = $row['nom'];
-        $prenom = $row['prenom'];
-        $urlImage = $row['urlImage'];
-        $role = $row['role'] ?? null;
-
-        // Retourner la Personnalite
-        return new Personnalite($id, $nom, $prenom, $urlImage, $role);
-    }
-
-    function hydrateAll(array $rows): array
-    {
-        $personnalites = [];
-        foreach ($rows as $row) {
-            $personnalite = $this->hydrate($row);
-            if ($personnalite) {
-                $personnalites[] = $personnalite;
-            }
-        }
-        return $personnalites;
-    }
-
     function find(int $id): ?Personnalite
     {
         $sql = "SELECT * FROM " . DB_PREFIX . "personnalite WHERE id = :id";
